wflow-jl/pyproject.toml
[project]
name = "wflow-hazard"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "dask>=2026.3.0",
    "geopandas>=1.1.3",
    "huggingface-hub>=1.14.0",
    "matplotlib>=3.10.9",
    "netcdf4>=1.7.4",
    "numpy>=2.4.4",
    "pandas>=3.0.3",
    "pyogrio>=0.12.1",
    "python-dotenv>=1.2.2",
    "requests>=2.34.2",
    "shapely>=2.1.2",
    "tqdm>=4.67.3",
    "xarray>=2026.4.0",
]
